Move judge protocol handler code to separate file
